More short -> int conversion.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1048 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/A6Record.java
@@ -26,7 +26,7 @@ private
 A6Record() {}
 
 private
-A6Record(Name name, short dclass, int ttl) {
+A6Record(Name name, int dclass, int ttl) {
 	super(name, Type.A6, dclass, ttl);
 }
 
@@ -42,7 +42,7 @@ getMember() {
  * @param prefix The name of the prefix
  */
 public
-A6Record(Name name, short dclass, int ttl, int prefixBits,
+A6Record(Name name, int dclass, int ttl, int prefixBits,
 	 Inet6Address suffix, Name prefix)
 {
 	this(name, dclass, ttl);
@@ -54,7 +54,7 @@ A6Record(Name name, short dclass, int ttl, int prefixBits,
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -75,7 +75,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	A6Record rec = new A6Record(name, dclass, ttl);
