+++ org/xbill/DNS/AAAARecord.java
@@ -24,7 +24,7 @@ private
 AAAARecord() {}
 
 private
-AAAARecord(Name name, short dclass, int ttl) {
+AAAARecord(Name name, int dclass, int ttl) {
 	super(name, Type.AAAA, dclass, ttl);
 }
 
@@ -38,13 +38,13 @@ getMember() {
  * @param address The address suffix
  */
 public
-AAAARecord(Name name, short dclass, int ttl, Inet6Address address) {
+AAAARecord(Name name, int dclass, int ttl, Inet6Address address) {
 	this(name, dclass, ttl);
 	this.address = address;
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -60,7 +60,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	AAAARecord rec = new AAAARecord(name, dclass, ttl);
