+++ org/xbill/DNS/HINFORecord.java
@@ -23,7 +23,7 @@ private
 HINFORecord() {}
 
 private
-HINFORecord(Name name, short dclass, int ttl) {
+HINFORecord(Name name, int dclass, int ttl) {
 	super(name, Type.HINFO, dclass, ttl);
 }
 
@@ -39,7 +39,7 @@ getMember() {
  * @throws IllegalArgumentException One of the strings has invalid escapes
  */
 public
-HINFORecord(Name name, short dclass, int ttl, String cpu, String os) {
+HINFORecord(Name name, int dclass, int ttl, String cpu, String os) {
 	this(name, dclass, ttl);
 	try {
 		this.cpu = byteArrayFromString(cpu);
@@ -51,7 +51,7 @@ HINFORecord(Name name, short dclass, int ttl, String cpu, String os) {
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -64,7 +64,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	HINFORecord rec = new HINFORecord(name, dclass, ttl);
