+++ org/xbill/DNS/Master.java
@@ -151,8 +151,7 @@ nextRecord() throws IOException {
 		if ((type = Type.value(s)) < 0)
 			throw st.exception("Invalid type '" + s + "'");
 
-		last = Record.fromString(name, (short) type, (short) dclass,
-					 ttl, st, origin);
+		last = Record.fromString(name, type, dclass, ttl, st, origin);
 		st.getEOL();
 		return last;
 	}
