+++ org/xbill/DNS/OPTRecord.java
@@ -32,7 +32,7 @@ private
 OPTRecord() {}
 
 private
-OPTRecord(Name name, short dclass, int ttl) {
+OPTRecord(Name name, int dclass, int ttl) {
 	super(name, Type.OPT, dclass, ttl);
 }
 
@@ -62,7 +62,7 @@ OPTRecord(short payloadSize, byte xrcode, byte version) {
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -84,7 +84,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	throw st.exception("no text format defined for OPT");
@@ -116,7 +116,7 @@ rdataToString() {
 }
 
 /** Returns the maximum allowed payload size. */
-public short
+public int
 getPayloadSize() {
 	return dclass;
 }
