+++ org/xbill/DNS/PTRRecord.java
@@ -22,7 +22,7 @@ private
 PTRRecord() {}
 
 private
-PTRRecord(Name name, short dclass, int ttl) {
+PTRRecord(Name name, int dclass, int ttl) {
 	super(name, Type.PTR, dclass, ttl);
 }
 
@@ -36,12 +36,12 @@ getMember() {
  * @param target The name of the machine with this address
  */
 public
-PTRRecord(Name name, short dclass, int ttl, Name target) {
+PTRRecord(Name name, int dclass, int ttl, Name target) {
 	super(name, Type.PTR, dclass, ttl, target);
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -49,7 +49,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	return rdataFromString(new PTRRecord(name, dclass, ttl), st, origin);
